Remove client-go dependency from library
