+++ pkg/v1/sdk/features/client/client.go
@@ -12,6 +12,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	configv1alpha1 "github.com/vmware-tanzu/tanzu-framework/apis/config/v1alpha1"
@@ -63,7 +64,7 @@ func getFeatureGateClient() (client.Client, error) {
 		return nil, err
 	}
 
-	if restConfig, err = config.GetCurrentClusterConfig(); err != nil {
+	if restConfig, err = getCurrentClusterConfig(); err != nil {
 		return nil, err
 	}
 
@@ -173,4 +174,26 @@ func (f *FeatureGateClient) setDeactivated(ctx context.Context, gate *configv1al
 		return fmt.Errorf("couldn't update featurgate %s: %w", gate.Name, err)
 	}
 	return nil
+}
+
+// getCurrentClusterConfig gets the config of current logged in cluster
+func getCurrentClusterConfig() (*rest.Config, error) {
+	server, err := config.GetCurrentServer()
+	if err != nil {
+		return nil, err
+	}
+	restConfig, err := getRestConfigWithContext(server.ManagementClusterOpts.Context, server.ManagementClusterOpts.Path)
+	if err != nil {
+		return nil, fmt.Errorf("couldn't get rest config: %w", err)
+	}
+	return restConfig, nil
+}
+
+// getRestConfigWithContext returns config using the passed context
+func getRestConfigWithContext(context, kubeconfigPath string) (*rest.Config, error) {
+	return clientcmd.NewNonInteractiveDeferredLoadingClientConfig(
+		&clientcmd.ClientConfigLoadingRules{ExplicitPath: kubeconfigPath},
+		&clientcmd.ConfigOverrides{
+			CurrentContext: context,
+		}).ClientConfig()
 }
